moved to epoll
the performances of epoll are better than poll
both for the polling elapsed time and for
the best api (no lookup is needed at the exit of the pool
as each event contains a pointer to the original structure).

Noteworthy the need to read and write until
errno is not equal to EAGAIN and EWOULDBLOCK
since the EDGE-TRIGGERED mode is used.

						printf("empty=%d\n", list_empty(&clients));